1. Under the snow

+++ 0x15-file_io/1-create_file.c
@@ -25,7 +25,7 @@ int create_file(const char *filename, char *text_content)
 		return (-1);
 
 	w = write(fd, text_content, len);
-	if (w == -1 || (size_t)w != len)
+	if (w == -1 || (size_t)w != (size_t)len)
 	{
 		close(fd);
 		return (-1);
